+++ qitos/kit/evaluate/__init__.py
@@ -3,11 +3,9 @@
 from .dsl_based import DSLEvaluator
 from .model_based import ModelBasedEvaluator
 from .rule_based import RuleBasedEvaluator
-from .cybench import CyBenchEvaluator
 
 __all__ = [
     "RuleBasedEvaluator",
     "DSLEvaluator",
     "ModelBasedEvaluator",
-    "CyBenchEvaluator",
 ]